move to shared place and fix
