+++ test/Extensions/TesterAzureUtils/Persistence/PersistenceGrainTests_AzureBlobStore_Json.cs
@@ -5,6 +5,7 @@ using Orleans.TestingHost;
 using TestExtensions;
 using Xunit;
 using Xunit.Abstractions;
+using TesterInternal.AzureInfra;
 
 namespace Tester.AzureUtils.Persistence
 {
