+++ test/Extensions/TesterAzureUtils/Persistence/PersistenceGrainTests_AzureStore.cs
@@ -18,6 +18,7 @@ using Orleans.Internal;
 using TestExtensions;
 using TestExtensions.Runners;
 using UnitTests.GrainInterfaces;
+using TesterInternal.AzureInfra;
 using AzureStoragePolicyOptions = Orleans.Clustering.AzureStorage.AzureStoragePolicyOptions;
 
 // ReSharper disable RedundantAssignment
