+++ test/Extensions/TesterAzureUtils/Persistence/PersistenceGrainTests_AzureTableGrainStorage.cs
@@ -9,6 +9,7 @@ using Orleans.Runtime.Configuration;
 using Orleans.Hosting;
 using Orleans.TestingHost;
 using TestExtensions;
+using TesterInternal.AzureInfra;
 
 namespace Tester.AzureUtils.Persistence
 {
