+++ test/Extensions/TesterAzureUtils/Streaming/AQSubscriptionObserverWithImplicitSubscribingTests.cs
@@ -7,6 +7,7 @@ using Orleans.Providers.Streams.AzureQueue;
 using Orleans.Streams;
 using Orleans.TestingHost;
 using Tester.StreamingTests.ProgrammaticSubscribeTests;
+using TesterInternal.AzureInfra;
 using TestExtensions;
 using Xunit;
 using Xunit.Abstractions;
