+++ test/Extensions/TesterAzureUtils/Streaming/SampleAzureQueueStreamingTests.cs
@@ -8,6 +8,7 @@ using Orleans.Hosting;
 using Orleans.Providers.Streams.AzureQueue;
 using Orleans.Runtime;
 using Orleans.TestingHost;
+using TesterInternal.AzureInfra;
 using TestExtensions;
 using UnitTests.StreamingTests;
 using Xunit;
